// whoxa_frontend/src/types/ResType.ts
export interface AllPublicGroup {
  group_profile_image: string;
  conversation_id: number;
  is_group: boolean;
  group_name: string;
  last_message: string;
  last_message_id: number;
  last_message_type: string;
  blocked_by_admin: boolean;
  created_by_admin: boolean;
  public_group: boolean;
  createdAt: Date;
  updatedAt: Date;
}